refactor(nav): make `<Switcher>` more reusable
